feat(tui): render the diff overlay on the source view (ADR 0046)
draw_source_screen now takes the session's parsed diff hunks and, when
the drilled-into symbol's file has an entry, composites
crate::source_diff::overlay_source_lines onto the rendered lines:
Added lines get ADDED_BG (winning over the symbol-range tint per ADR
0046 decision 6), Removed lines render as extra "-"-gutter rows with
REMOVED_BG, and drift between the file on disk and the diff drops the
overlay with a note in the pane title instead of a wrong mapping.

crate::ui::draw grows a diff_hunks parameter threaded from
crate::run_app (#[allow(too_many_arguments)], matching the existing
precedent in blast_radius.rs) — every non-source-screen call site
passes it through unused, only draw_source_screen consumes it.

Splits source_screen.rs's tests into source_screen/tests.rs (file-size
discipline, same pattern as the source_diff split).
