+++ rinkaku-tui/src/ui/detail_pane_tests/content_by_row_kind.rs
@@ -33,6 +33,7 @@ fn should_draw_placeholder_message_when_there_are_no_rows_at_all() {
                 &[],
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -79,6 +80,7 @@ fn should_draw_dir_detail_content_when_cursor_is_on_a_directory_row() {
                 &[],
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -130,6 +132,7 @@ fn should_draw_symbols_label_without_changed_wording_when_origin_is_repo_outline
                 &[],
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -163,6 +166,7 @@ fn should_draw_skip_reason_in_detail_pane_when_cursor_is_on_a_skipped_file_row()
                 &[],
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -194,6 +198,7 @@ fn should_draw_test_symbol_count_in_detail_pane_when_cursor_is_on_a_whole_test_f
                 &[],
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -253,6 +258,7 @@ fn should_draw_both_test_note_and_real_symbols_in_detail_pane_when_file_is_mixed
                 &[],
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -284,6 +290,7 @@ fn should_draw_detail_pane_content_when_cursor_is_on_a_symbol_row() {
                 &[],
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
