+++ rinkaku-tui/src/ui/diff_pane_tests/styling.rs
@@ -31,6 +31,7 @@ index e69de29..4b825dc 100644
                 &diff_highlights,
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -77,6 +78,7 @@ index e69de29..4b825dc 100644
                 &diff_highlights,
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -116,6 +118,7 @@ index e69de29..4b825dc 100644
                 &diff_highlights,
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -162,6 +165,7 @@ index e69de29..4b825dc 100644
                 &diff_highlights,
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
@@ -226,6 +230,7 @@ index e69de29..4b825dc 100644
                 &diff_highlights,
                 &BlastRadiusSelection::NotApplicable,
                 None,
+                &[],
             );
         })
         .expect("draw");
