+++ rinkaku-tui/src/ui/entry.rs
@@ -239,6 +239,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
@@ -286,6 +287,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
@@ -315,6 +317,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
@@ -354,6 +357,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
@@ -408,6 +412,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
@@ -468,6 +473,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
@@ -499,6 +505,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
@@ -531,6 +538,7 @@ mod tests {
                     &[],
                     &BlastRadiusSelection::NotApplicable,
                     None,
+                    &[],
                 );
             })
             .expect("draw");
